+++ packages/DatePicker/package.json
@@ -34,10 +34,10 @@
   },
   "homepage": "https://github.com/WTTJ/welcome-ui#readme",
   "dependencies": {
-    "@welcome-ui/date-time-picker": "^1.6.2",
     "@welcome-ui/icons.date": "^1.5.2",
     "@welcome-ui/system": "^1.5.2",
     "@welcome-ui/utils": "^1.4.0",
+    "@welcome-ui/date-time-picker-common": "^1.6.2",
     "react-datepicker": "^2.9.6"
   },
   "peerDependencies": {
