--- packages/DatePicker/styles.js
@@ -1,9 +1,0 @@
-import styled from '@xstyled/styled-components'
-import ReactDatePicker from 'react-datepicker'
-import { componentSystem } from '@welcome-ui/system'
-import { fieldStyles } from '@welcome-ui/utils'
-
-export const DatePicker = styled(ReactDatePicker)`
-  ${fieldStyles};
-  ${componentSystem};
-`
