+++ packages/DateTimePicker/package.json
@@ -43,7 +43,7 @@
     "@welcome-ui/select": "^1.6.2",
     "@welcome-ui/system": "^1.5.2",
     "@welcome-ui/time-picker": "^1.6.2",
-    "lodash.range": "^3.2.0",
+    "@welcome-ui/date-time-picker-common": "^1.6.2",
     "react-datepicker": "^2.9.6"
   },
   "peerDependencies": {
