+++ packages/DateTimePickerCommon/package.json
@@ -1,9 +1,9 @@
 {
-  "name": "@welcome-ui/time-picker",
+  "name": "@welcome-ui/date-time-picker-common",
   "version": "1.6.2",
   "description": "Customizable design system with react • styled-components • styled-system and reakit.",
-  "main": "dist/time-picker.cjs.js",
-  "module": "dist/time-picker.es.js",
+  "main": "dist/date-time-picker-common.cjs.js",
+  "module": "dist/date-time-picker-common.es.js",
   "scripts": {
     "test": "jest"
   },
@@ -34,20 +34,24 @@
   },
   "homepage": "https://github.com/WTTJ/welcome-ui#readme",
   "dependencies": {
-    "@welcome-ui/system": "^1.5.2",
-    "@welcome-ui/utils": "^1.4.0",
-    "@welcome-ui/date-time-picker-common": "^1.6.2",
-    "react-datepicker": "^2.9.6"
+    "lodash.range": "^3.2.0",
+    "@welcome-ui/button": "^1.5.2",
+    "@welcome-ui/clear-button": "^1.6.0",
+    "@welcome-ui/field": "^1.5.2",
+    "@welcome-ui/icons.left": "^1.5.2",
+    "@welcome-ui/icons.right": "^1.5.2",
+    "@welcome-ui/select": "^1.6.2",
+    "@welcome-ui/system": "^1.5.2"
   },
   "peerDependencies": {
     "@xstyled/styled-components": "^1.15.1",
+    "@xstyled/system": "^1.15.1",
     "prop-types": "^15.7.2",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
-    "styled-components": "^4.0.0 || ^5.0.0"
+    "styled-components": "^5.0.0"
   },
   "gitHead": "5e05d638c444651c818138903aef4a68cbbb5747",
   "sideEffects": false,
-  "component": "TimePicker",
-  "documentation": "http://welcome-ui.com/fields/time-picker"
+  "component": "DateTimePickerCommon"
 }
